ai/langchain/requirements.txt
python-dotenv
openai
wikipedia
langchain-core 
langchain-community 
langchain-openai
langchain-cli
langchain-fireworks
numexpr